+++ lib/managers/ar_session_manager.dart
@@ -1,8 +1,14 @@
+import 'dart:math' show sqrt;
 import 'dart:typed_data';
+
+import 'package:ar_flutter_plugin/datatypes/config_planedetection.dart';
+import 'package:ar_flutter_plugin/models/ar_anchor.dart';
 import 'package:ar_flutter_plugin/models/ar_hittest_result.dart';
+import 'package:ar_flutter_plugin/utils/json_converters.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:ar_flutter_plugin/datatypes/config_planedetection.dart';
+import 'package:vector_math/vector_math_64.dart';
+import 'package:vector_math/vector_math_64.dart' as vecmath;
 
 // Type definitions to enforce a consistent use of the API
 typedef ARHitResultHandler = void Function(List<ARHitTestResult> hits);
@@ -11,6 +17,8 @@ typedef ARHitResultHandler = void Function(List<ARHitTestResult> hits);
 class ARSessionManager {
   /// Platform channel used for communication from and to [ARSessionManager]
   late MethodChannel _channel;
+
+  MethodChannel get channel => _channel;
 
   /// Debugging status flag. If true, all platform calls are printed. Defaults to false.
   final bool debug;
@@ -31,6 +39,71 @@ class ARSessionManager {
     if (debug) {
       print("ARSessionManager initialized");
     }
+  }
+
+  /// Returns the camera pose in Matrix4 format with respect to the world coordinate system of the [ARView]
+  Future<Matrix4?> getCameraPose() async {
+    try {
+      final serializedCameraPose =
+          await _channel.invokeMethod<List<dynamic>>('getCameraPose', {});
+      return MatrixConverter().fromJson(serializedCameraPose!);
+    } catch (e) {
+      print('Error caught: ' + e.toString());
+      return null;
+    }
+  }
+
+  /// Returns the given anchor pose in Matrix4 format with respect to the world coordinate system of the [ARView]
+  Future<Matrix4?> getPose(ARAnchor anchor) async {
+    try {
+      if (anchor.name.isEmpty) {
+        throw Exception("Anchor can not be resolved. Anchor name is empty.");
+      }
+      final serializedCameraPose =
+          await _channel.invokeMethod<List<dynamic>>('getAnchorPose', {
+        "anchorId": anchor.name,
+      });
+      return MatrixConverter().fromJson(serializedCameraPose!);
+    } catch (e) {
+      print('Error caught: ' + e.toString());
+      return null;
+    }
+  }
+
+  /// Returns the distance in meters between @anchor1 and @anchor2.
+  Future<double?> getDistanceBetweenAnchors(
+      ARAnchor anchor1, ARAnchor anchor2) async {
+    var anchor1Pose = await getPose(anchor1);
+    var anchor2Pose = await getPose(anchor2);
+    var anchor1Translation = anchor1Pose?.getTranslation();
+    var anchor2Translation = anchor2Pose?.getTranslation();
+    if (anchor1Translation != null && anchor2Translation != null) {
+      return getDistanceBetweenVectors(anchor1Translation, anchor2Translation);
+    } else {
+      return null;
+    }
+  }
+
+  /// Returns the distance in meters between @anchor and device's camera.
+  Future<double?> getDistanceFromAnchor(ARAnchor anchor) async {
+    Matrix4? cameraPose = await getCameraPose();
+    Matrix4? anchorPose = await getPose(anchor);
+    Vector3? cameraTranslation = cameraPose?.getTranslation();
+    Vector3? anchorTranslation = anchorPose?.getTranslation();
+    if (anchorTranslation != null && cameraTranslation != null) {
+      return getDistanceBetweenVectors(anchorTranslation, cameraTranslation);
+    } else {
+      return null;
+    }
+  }
+
+  /// Returns the distance in meters between @vector1 and @vector2.
+  double getDistanceBetweenVectors(Vector3 vector1, Vector3 vector2) {
+    num dx = vector1.x - vector2.x;
+    num dy = vector1.y - vector2.y;
+    num dz = vector1.z - vector2.z;
+    double distance = sqrt(dx * dx + dy * dy + dz * dz);
+    return distance;
   }
 
   Future<void> _platformCallHandler(MethodCall call) {
@@ -122,5 +195,98 @@ class ARSessionManager {
   Future<ImageProvider> snapshot() async {
     final result = await _channel.invokeMethod<Uint8List>('snapshot');
     return MemoryImage(result!);
+  }
+
+  int _logoIndex = -1;
+
+  // TODO : transform this as to return a quaternion instead of a vector4
+  Future<vecmath.Vector4> getCameraOrientation() async {
+    vecmath.Vector4 cameraOrientation = vecmath.Vector4.zero();
+    var data = await channel.invokeMethod<Float32List>('getCameraOrientation');
+
+    cameraOrientation.x = data?.elementAt(0) ?? 0;
+    cameraOrientation.y = data?.elementAt(1) ?? 0;
+    cameraOrientation.z = data?.elementAt(2) ?? 0;
+    cameraOrientation.w = data?.elementAt(3) ?? 1;
+
+    return cameraOrientation;
+  }
+
+  Future<String> startRecording(String path) async {
+    String file_path =
+        await channel.invokeMethod<String>('startRecording', {'path': path}) ??
+            'error';
+    return file_path;
+  }
+
+  Future<void> stopRecording() async {
+    await channel.invokeMethod<void>('stopRecording');
+  }
+
+  Future<int> getRecordLength(String path) async {
+    int duration =
+        await channel.invokeMethod<int>('getRecordLength', {'path': path}) ??
+            -1;
+    return duration;
+  }
+
+  Future<bool> enableLogoTracking() async {
+    _logoIndex = await channel.invokeMethod<int>('enableLogoTracking') ?? -1;
+
+    print('logo index : ' + _logoIndex.toString());
+
+    return true;
+  }
+
+  Future<vecmath.Vector4> getQuaternionFromLogo() async {
+    vecmath.Vector4 calibrationQuaternion = vecmath.Vector4.zero();
+    var data = await channel.invokeMethod<Float32List>(
+        'getQuaternionFromLogo', {'index': _logoIndex});
+
+    calibrationQuaternion.x = data?.elementAt(0) ?? 0;
+    calibrationQuaternion.y = data?.elementAt(1) ?? 0;
+    calibrationQuaternion.z = data?.elementAt(2) ?? 0;
+    calibrationQuaternion.w = data?.elementAt(3) ?? 1;
+
+    return calibrationQuaternion;
+  }
+
+  Future<vecmath.Vector3> getTranslationFromLogo() async {
+    vecmath.Vector3 translationVector = vecmath.Vector3.zero();
+    var data = await channel.invokeMethod<Float32List>(
+        'getTranslationFromLogo', {'index': _logoIndex});
+
+    translationVector.x = data?.elementAt(0) ?? 0;
+    translationVector.y = data?.elementAt(1) ?? 0;
+    translationVector.z = data?.elementAt(2) ?? 0;
+
+    return translationVector;
+  }
+
+  Future<bool> isLogoTracked() async {
+    if (_logoIndex == -1) return false;
+
+    var data = await channel
+        .invokeMethod<bool>('isLogoTracked', {'index': _logoIndex});
+
+    bool result = data ?? false;
+
+    return result;
+  }
+
+  Future<String> getVIOTrackingStatus() async {
+    var data = await channel.invokeMethod<String>('getTrackingStatus');
+
+    String result = data ?? "";
+
+    return result;
+  }
+
+  Future<String> getVIOTrackingError() async {
+    var data = await channel.invokeMethod<String>('getTrackingError');
+
+    String result = data ?? "";
+
+    return result;
   }
 }